Add files via upload
Senior project involves web application with known vulnerabilities... Given that this website was created on an ancient framework(JSP) and I know its vulnerable (Oops) I know this is going to be a good project to use. It is important to note that I rewrote a schema and it is in the project as well as some mock data and I even fixed some old bugs that were really bothering me!