Add files via upload
Reuploading this project since it is going to be used in my senior project. My senior project is looking at vulnerable legacy web applications and this was built on an ancient framework (JSP) this is a perfect choice. I reverse engineered the database and have put the schema in the project as well as some mock data. This is still a highly vulnerable web app (oops) and I even fixed some bugs that were bothering me from 2 years ago!